refactor(button): create button base component that supports colors
DSY-1460
