+++ src/components/Button/Button.tsx
@@ -1,45 +1,8 @@
 import React from 'react';
-import styled, { withTheme } from 'styled-components/native';
 import { NativeSyntheticEvent, NativeTouchEvent } from 'react-native';
-import {
-  getColorPrimary,
-  getColorOnPrimary,
-  getColorHighEmphasis,
-  getColorMediumEmphasis,
-  getColorLowEmphasis,
-  getButtonPropsBySize,
-  getRadiusBySize,
-  getShadowBySize,
-  Theme,
-  getOpacity10,
-  getColorPrimaryLight,
-} from '../../common/themeSelectors';
-
-export type ButtonTypes = 'contained' | 'outlined' | 'text'
+import { ButtonBase, ButtonTypes } from './ButtonBase';
 
 export interface ButtonProps {
-  /**
-   * The onPress event handler
-   */
-  onPress: (ev: NativeSyntheticEvent<NativeTouchEvent>) => void,
-  /**
-   * The button text content
-   */
-  text: string
-  /**
-   * Button variants `contained` | `outlined` | `text`
-   */
-  type?: ButtonTypes
-  /**
-   * A disabled button is unusable and un-clickable.
-   * The disabled attribute can be set to keep a user from clicking on the button until some
-   * other condition has been met (like selecting a checkbox, etc.).
-   */
-  disabled?: boolean
-  /**
-   * The button theme
-   */
-  theme: Theme,
   /**
    * An accessibility hint helps users understand what will happen when they perform an action
    * on the accessibility element when that result is not clear from the accessibility label.
@@ -52,82 +15,46 @@ export interface ButtonProps {
    */
   accessibilityLabel?: string
   /**
+   * A disabled button is unusable and un-clickable.
+   * The disabled attribute can be set to keep a user from clicking on the button until some
+   * other condition has been met (like selecting a checkbox, etc.).
+   */
+  disabled?: boolean
+  /**
+   * The onPress event handler
+   */
+  onPress: (ev: NativeSyntheticEvent<NativeTouchEvent>) => void,
+  /**
   * Optional ID for testing
   */
   testID?: string,
+  /**
+   * The button text content
+   */
+  text: string
+  /**
+   * Button variants `contained` | `outlined` | `text`
+   */
+  type?: ButtonTypes
 }
 
-interface ButtonBase {
-  type: ButtonTypes
-  disabled: boolean
-  theme: Theme
-}
-
-const isContained = (type: ButtonTypes) => type === 'contained';
-
-const getButtonStyles = (theme: Theme, type: ButtonTypes, disabled: boolean) => {
-  const styles = {
-    contained: {
-      background: disabled ? getColorLowEmphasis(theme) : getColorPrimary(theme),
-    },
-    outlined: {
-      borderColor: disabled ? getColorMediumEmphasis(theme) : getColorPrimary(theme),
-      borderWidth: 1,
-    },
-  };
-
-  return styles[type];
-};
-
-const getButtonTextColor = (theme: Theme, type: ButtonTypes, disabled: boolean) => {
-  const color = {
-    active: isContained(type) ? getColorOnPrimary(theme) : getColorHighEmphasis(theme),
-    disabled: isContained(type) ? getColorHighEmphasis(theme) : getColorMediumEmphasis(theme),
-  };
-
-  return disabled ? color.disabled : color.active;
-};
-
-const ButtonBase = styled.TouchableHighlight<ButtonBase>(({ type, theme, disabled = false }) => ({
-  borderRadius: getRadiusBySize(theme, 'medium'),
-  ...getButtonStyles(theme, type, disabled),
-  ...getButtonPropsBySize(theme, 'medium'),
-}));
-
-const Text = styled.Text<ButtonBase>`
-  color: ${({ type, theme, disabled }) => getButtonTextColor(theme, type, disabled)};
-  font-size: 14px;
-  align-self: center;
-  letter-spacing: 1px;
-`;
-
-const getShadowByType = (type: ButtonTypes, disabled: boolean, theme: Theme) => (
-  isContained(type) && !disabled
-    ? getShadowBySize(theme, '2')
-    : {}
-);
-
-const ButtonComponent = ({
-  onPress, theme, text, type = 'contained', disabled = false, testID = 'button', accessibilityLabel, accessibilityHint,
+export const Button = ({
+  accessibilityHint,
+  accessibilityLabel,
+  disabled = false,
+  onPress,
+  testID = 'button',
+  text,
+  type = 'contained',
 }: ButtonProps) => (
   <ButtonBase
-    testID={testID}
-    type={type}
-    onPress={disabled ? () => {} : onPress}
-    style={getShadowByType(type, disabled, theme)}
-    underlayColor={getColorPrimaryLight(theme)}
-    activeOpacity={getOpacity10(theme)}
-    disabled={disabled}
-  >
-    <Text
-      style={{ fontWeight: 'bold' }}
-      accessibilityLabel={accessibilityLabel}
-      accessibilityHint={accessibilityHint}
-      accessibilityRole="button"
-      type={type}
-      disabled={disabled}
-    >{text.toUpperCase()}</Text>
-  </ButtonBase>
+    accessibilityHint={ accessibilityHint }
+    accessibilityLabel={ accessibilityLabel }
+    disabled={ disabled }
+    textColor='default'
+    onPress={ onPress }
+    testID={ testID }
+    text={ text }
+    type={ type }
+  />
 );
-
-export const Button = withTheme(ButtonComponent);
